refactor(tonk-render): extract page orchestrator behind a query trait
Move slide's model -> view -> entity -> HTML orchestration out of the
binary and into tonk-render's new `page` module, behind a `QueryBackend`
trait (the one host-specific seam: a ConceptQuery to a Vec<Conclusion>).
slide keeps its public render surface and now implements the trait over
its reactor; a worker SSR route can reuse the same orchestrator by
implementing the trait over a URL-named branch. Pure renderer and page
orchestrator share one crate; tonk-template stays the planner both the
browser and headless paths compute against.
